Added a function pointer to amplitude and frequency arrays to allow reloading of data.
Added a save file container for containing all the EnvelopeMakers.

Added pairs to the function class to allow reloading of classes. Need to update function classes to store points.

Added native methods to pass data to Java for drawing on a bare View class that needs to be fleshed out.