Added a base Envelope class to hold the functions and forward getMax and getMin calls to the functions.
WaveMaker now gets the max and min of all the envelopes it holds.

Created test code yet to be run. This includes the FunctionView which should be able to draw function data to it's canvas now.

Moved all the native methods to their own Java class.

Created an empty main layout where the main GUI actions are.